Added mongodb

+++ src/blogs/routers/handlers/delete-blog-handler.ts
@@ -3,14 +3,16 @@ import { blogsRepository } from '../../repositories/blogs-repository';
 import { HTTP_STATUSES } from '../../../core/types/http-statuses';
 import { CreateErrorMessage } from '../../../core/utils/error-utils';
 
-export const deleteBlogHandler = (
+export const deleteBlogHandler = async (
   req: Request<{ id: string }>,
   res: Response,
 ) => {
-  const isDeleted = blogsRepository.delete(req.params.id);
+  const blog = await blogsRepository.findById(req.params.id);
 
-  if (isDeleted) {
+  if (blog) {
+    await blogsRepository.delete(req.params.id);
     res.sendStatus(HTTP_STATUSES.NO_CONTENT_204);
+    return;
   }
 
   res
